started writing a map loader (pending details on map json format from alch). Switched file loading from absolute path to Mod.GetFileStream on class/map loading. changed class pick to an enum and eliminated switches in classsystem/commandsystem. Changed class.json format to contain hp/mana data (CLASS JSON REFORMAT NEEDED, MATCH TO archer.json FORMAT)

+++ Content/Classes/Abilities.cs
@@ -454,7 +454,7 @@ namespace CTG2.Content
 
             if (Player.HeldItem.type == ItemID.WhoopieCushion && Player.controlUseItem && Player.itemTime == 0 && !Player.HasBuff(BuffID.ChaosState)) // Only activate if not on cooldown
             {
-                int selectedClass = Player.GetModPlayer<ClassSystem>().playerClass;
+                int selectedClass = (int)Player.GetModPlayer<ClassSystem>().playerClass;
 
                 switch (selectedClass)
                 {
